refactor: remove onComplete and AI image generation functionality from DreamBoard, simplifying component logic and enhancing clarity
